probe: extract Lightweight probe to internal/probe and module/probe
Lightweight's only Server dependency was *client.ClientPool, which makes
it the smallest of the three strategies to decouple.

Service layer (internal/probe/lightweight.go):
  - LightweightService struct with *client.ClientPool injected via
    NewLightweightService.
  - Public surface: a single Probe(ctx, *typ.Provider)
    *LightweightProbeResponseData that never errors — partial failure is
    encoded in the per-endpoint fields and the Valid summary, matching
    the original behaviour where /probe/lightweight is informational
    only and must not block onboarding.
  - The four endpoint helpers (OPTIONS, /models, /chat, /responses)
    moved over with their logic unchanged. The four inline anonymous
    return structs in the original were collapsed to two named local
    types (endpointReport, modelsReport) — pure readability win, no
    behaviour change.

HTTP layer (internal/server/module/probe):
  - Handler grows a second field (lightweight) and a LightweightResponse
    envelope with the same local errorDetail used by V2.
  - HandleLightweightProbe owns request decoding + the
    api_base/api_style/token validation that previously lived in
    HandleLightweightProbe on *Server, then delegates to the service.
  - RegisterRoutes adds POST /probe/lightweight with identical swagger
    annotations.

Server wiring:
  - New field probeLightweight on *Server, constructed alongside
    probeV2Service in NewServer.
  - NewHandler signature extended; webui.go drops the direct
    s.HandleLightweightProbe registration in favour of the module call.
  - LightweightProbeResponse wrapper deleted from server_types.go — no
    longer referenced.

Cleanup:
  - internal/server/lightweight_probe.go deleted (444 LOC).
  - Removed the now-unused probe import from webui.go.

Net change: server -456 LOC, probe +254 LOC, module/probe +79 LOC.
With Lightweight done, the only probe strategy still living in
internal/server is Adaptive, which is genuinely coupled via probeCache,
capabilityStore, adaptive_endpoint_router, and the health monitor —
that's a separate, larger piece of work.

Build/vet/test green.
