+++ internal/server/server.go
@@ -117,6 +117,9 @@ type Server struct {
 	// probeV2Service runs SDK-level end-to-end probes for the /api/v2/probe endpoint.
 	probeV2Service *probe.V2Service
 
+	// probeLightweight powers /api/v2/probe/lightweight — optional key validation.
+	probeLightweight *probe.LightweightService
+
 	// mcp runtime for external MCP tools
 	mcpRuntime *mcpruntime.Runtime
 
@@ -757,6 +760,7 @@ func NewServer(cfg *config.Config, opts ...ServerOption) *Server {
 	// V2 probe service handles /api/v2/probe end-to-end without touching *Server.
 	// The smart-routing callback closes over the server so probe doesn't import server.
 	server.probeV2Service = probe.NewV2Service(cfg, server.clientPool, server.SelectServiceFromSmartRouting)
+	server.probeLightweight = probe.NewLightweightService(server.clientPool)
 
 	// Initialize OTel meter setup for token tracking
 	sm := cfg.StoreManager()
